refactored results into separate function
