--- src/components/Table.jsx
@@ -1,24 +1,0 @@
-export default function Table() {
-  return (
-    <table id="result">
-      <thead>
-        <tr>
-          <th>Year</th>
-          <th>Investment Value</th>
-          <th>Interest (Year)</th>
-          <th>Total Interest</th>
-          <th>Investment Capital</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td>data</td>
-          <td>data</td>
-          <td>data</td>
-          <td>data</td>
-          <td>data</td>
-        </tr>
-      </tbody>
-    </table>
-  );
-}
